fix: resolve circular dependency and implement AU update logic
- Created `src/logics/exrStateLogic.ts` to house `getUpdatedExRState`, breaking the circular dependency between slices, syncLogic, and useStore.
- Completed the implementation of AU option updates with proper ExR state synchronization.
- Verified that all tests pass and Biome check is successful.
