+++ src/slices/auOptionViewerSlice.ts
@@ -1,7 +1,7 @@
 import type { StateCreator } from "zustand";
 import { auOptionMetaData, updateAuOption } from "../logics/api";
 import { parseAuOptionId } from "../logics/optionUtils";
-import { getUpdatedExRState } from "../logics/syncLogic";
+import { getUpdatedExRState } from "../logics/exrStateLogic";
 import type { AuOptionId, AuRoleOption } from "../type";
 import { OptionValueType } from "../type";
 import type { ExROptionViewerSlice } from "./exrOptionViewerSlice";
